pyproject.toml
[project]
name = "risk-matrix-scheduler"
version = "0.1.0"
description = ""
authors = [
    {name = "perdanamain",email = "firman.fp123@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "schedule (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
